+++ crates/net/network/src/peers/mod.rs
@@ -1,0 +1,8 @@
+//! Peer related implementations
+
+mod manager;
+mod reputation;
+
+pub(crate) use manager::{PeerAction, PeersManager};
+pub use manager::{PeersConfig, PeersHandle};
+pub(crate) use reputation::ReputationChange;
